fw att ctrl: airspeed check: remove unnecessary cast

+++ src/modules/fw_att_control/fw_att_control_main.cpp
@@ -633,7 +633,7 @@ FixedwingAttitudeControl::task_main()
 				/* if airspeed is smaller than min, the sensor is not giving good readings */
 				if ((_airspeed.indicated_airspeed_m_s < 0.5f * _parameters.airspeed_min) ||
 				    !isfinite(_airspeed.indicated_airspeed_m_s) ||
-				    (float)hrt_elapsed_time(&_airspeed.timestamp) > 1e6f) {
+				    hrt_elapsed_time(&_airspeed.timestamp) > 1e6) {
 					airspeed = _parameters.airspeed_trim;
 
 				} else {
